refactor: move compat.ModNameCompat classes to compat.modname.ModNameCompat
