+++ src/main/java/yesman/epicfight/compat/playeranimator/PlayerAnimatorCompat.java
@@ -1,11 +1,14 @@
-package yesman.epicfight.compat;
+package yesman.epicfight.compat.playeranimator;
 
+import dev.kosmx.playerAnim.impl.IAnimatedPlayer;
+import dev.kosmx.playerAnim.impl.animation.AnimationApplier;
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.api.distmarker.OnlyIn;
 import net.neoforged.bus.api.IEventBus;
 import yesman.epicfight.api.client.neoevent.RenderEpicFightPlayerEvent;
+import yesman.epicfight.compat.ICompatModule;
 
-public class MCreatorPlayerAnimationsCompat implements ICompatModule {
+public class PlayerAnimatorCompat implements ICompatModule {
     @Override
     public void onModEventBus(IEventBus eventBus) {}
 
@@ -23,7 +26,7 @@ public class MCreatorPlayerAnimationsCompat implements ICompatModule {
 
     @OnlyIn(Dist.CLIENT)
     private void renderEvent(RenderEpicFightPlayerEvent event) {
-        String animation = event.getPlayerPatch().getOriginal().getPersistentData().getString("PlayerCurrentAnimation");
-        if (!animation.isEmpty()) event.setShouldRender(false);
+        AnimationApplier emote = ((IAnimatedPlayer) event.getPlayerPatch().getOriginal()).playerAnimator_getAnimation();
+        if (emote.isActive()) event.setShouldRender(false);
     }
 }
