+++ src/main/java/yesman/epicfight/compat/werewolves/WerewolvesCompat.java
@@ -1,4 +1,4 @@
-package yesman.epicfight.compat;
+package yesman.epicfight.compat.werewolves;
 
 import com.mojang.blaze3d.vertex.PoseStack;
 
@@ -28,6 +28,7 @@ import yesman.epicfight.client.renderer.EpicFightRenderTypes;
 import yesman.epicfight.client.renderer.patched.entity.PPlayerRenderer;
 import yesman.epicfight.client.renderer.patched.layer.PatchedLayer;
 import yesman.epicfight.client.world.capabilites.entitypatch.player.AbstractClientPlayerPatch;
+import yesman.epicfight.compat.ICompatModule;
 import yesman.epicfight.mixin.teamlapen.MixinHumanWerewolfLayer;
 
 public class WerewolvesCompat implements ICompatModule {
