+++ src/main/java/yesman/epicfight/main/EpicFightMod.java
@@ -52,6 +52,17 @@ import yesman.epicfight.client.gui.screen.config.ItemsPreferenceScreen;
 import yesman.epicfight.client.renderer.patched.item.EpicFightItemProperties;
 import yesman.epicfight.client.renderer.shader.compute.loader.ComputeShaderProvider;
 import yesman.epicfight.compat.*;
+import yesman.epicfight.compat.azurelib.AzureLibArmorCompat;
+import yesman.epicfight.compat.azurelib.AzureLibCompat;
+import yesman.epicfight.compat.curiosapi.CuriosCompat;
+import yesman.epicfight.compat.firstperson.FirstPersonCompat;
+import yesman.epicfight.compat.geckolib.GeckolibCompat;
+import yesman.epicfight.compat.iris.IRISCompat;
+import yesman.epicfight.compat.mcreator.MCreatorPlayerAnimationsCompat;
+import yesman.epicfight.compat.playeranimator.PlayerAnimatorCompat;
+import yesman.epicfight.compat.skinlayer3d.SkinLayer3DCompat;
+import yesman.epicfight.compat.vampirism.VampirismCompat;
+import yesman.epicfight.compat.werewolves.WerewolvesCompat;
 import yesman.epicfight.config.ClientConfig;
 import yesman.epicfight.config.CommonConfig;
 import yesman.epicfight.config.ServerConfig;
